### 2022-10-17 #4860 @editorialbot
[PRE REVIEW]: hstrat: a Python Package for phylogenetic inference on distributed digital evolution populations
**Submitting author:** @mmore500 (<a href="http://orcid.org/0000-0003-4726-4479">Matthew Andres Moreno</a>)
**Repository:** https://github.com/mmore500/hstrat
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.3
