### 2022-10-17 #4860 @danielskatz
@mmore500 - thanks for your submission to JOSS. I'll be looking for an editor for this.

While I do so, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Also, if you have references in your bib file that you are not actually citing in the paper, you could remove them, as our bot checks everything in the bib file, whether it is used in the paper or not.
