### 2022-10-18 8eca0cf
Fixup citations
Crawl publisher websites for bibtex
