+++ joss/paper.bib
@@ -1,56 +1,79 @@
 @inproceedings{ackley2011homeostatic,
-  title = {Homeostatic architectures for robust spatial computing},
-  author = {Ackley, David H and Williams, Lance R},
+  title = {Homeostatic Architectures for Robust Spatial Computing},
+  author = {Ackley, David H. and Williams, Lance R.},
   year = 2011,
   booktitle = {2011 Fifth IEEE Conference on Self-Adaptive and Self-Organizing Systems Workshops},
   pages = {91--96},
-  organization = {Ieee},
+  doi = {10.1109/sasow.2011.18},
 }
 @inproceedings{ackley2011pursue,
-  title = {Pursue robust indefinite scalability},
-  author = {Ackley, David H and Cannon, Daniel C},
+  title = {Pursue Robust Indefinite Scalability},
+  author = {David H. Ackley and Daniel C. Cannon},
   year = 2011,
+  month = may,
   booktitle = {13th Workshop on Hot Topics in Operating Systems (HotOS XIII)},
+  publisher = {USENIX Association},
+  address = {Napa, CA},
+  url = {https://www.usenix.org/conference/hotosxiii/pursue-robust-indefinite-scalability},
 }
 @inproceedings{ackley2014indefinitely,
-  title = {Indefinitely scalable computing= artificial life engineering},
+  title = {Indefinitely Scalable Computing = Artificial Life Engineering},
   author = {Ackley, David and Small, Trent},
   year = 2014,
-  booktitle = {ALIFE 14: The Fourteenth International Conference on the Synthesis and Simulation of Living Systems},
+  month = {07},
+  series = {ALIFE 2022: The 2022 Conference on Artificial Life},
+  volume = {ALIFE 14: The Fourteenth International Conference on the Synthesis and Simulation of Living Systems},
   pages = {606--613},
-  organization = {MIT Press},
+  doi = {10.1162/978-0-262-32621-6-ch098},
+  url = {https://doi.org/10.1162/978-0-262-32621-6-ch098},
+  eprint = {https://direct.mit.edu/isal/proceedings-pdf/alife2014/26/606/1901859/978-0-262-32621-6-ch098.pdf},
 }
 @inproceedings{angluin2005power,
-  title = {On the power of anonymous one-way communication},
+  title = {On the Power of Anonymous One-Way Communication},
   author = {Angluin, Dana and Aspnes, James and Eisenstat, David and Ruppert, Eric},
-  year = 2005,
-  booktitle = {International Conference On Principles Of Distributed Systems},
+  year = 2006,
+  booktitle = {Principles of Distributed Systems},
+  publisher = {Springer Berlin Heidelberg},
+  address = {Berlin, Heidelberg},
   pages = {396--411},
-  organization = {Springer},
+  doi = {10.1007/11795490_30},
+  isbn = {978-3-540-36322-4},
+  editor = {Anderson, James H. and Prencipe, Giuseppe and Wattenhofer, Roger},
 }
 @article{aspnes2009introduction,
-  title = {An introduction to population protocols},
+  title = {An Introduction to Population Protocols},
   author = {Aspnes, James and Ruppert, Eric},
   year = 2009,
-  journal = {Middleware for Network Eccentric and Mobile Applications},
-  publisher = {Springer},
+  booktitle = {Middleware for Network Eccentric and Mobile Applications},
+  publisher = {Springer Berlin Heidelberg},
+  address = {Berlin, Heidelberg},
   pages = {97--120},
+  doi = {10.1007/978-3-540-89707-1_5},
+  isbn = {978-3-540-89707-1},
[…]
