### 2022-10-18 d2088e6
Remove null fields
