+++ joss/paper.bib
@@ -409,9 +409,6 @@
   month = 10,
   journal = {PLOS Biology},
   publisher = {Public Library of Science},
-  volume = 1,
-  number = 1,
-  pages = {null},
   doi = {10.1371/journal.pbio.0000018},
   url = {https://doi.org/10.1371/journal.pbio.0000018},
 }
