+++ joss/paper.bib
@@ -16,6 +16,7 @@
   address = {Napa, CA},
   url = {https://www.usenix.org/conference/hotosxiii/pursue-robust-indefinite-scalability},
 }
+% dead doi
 @inproceedings{ackley2014indefinitely,
   title = {Indefinitely Scalable Computing = Artificial Life Engineering},
   author = {Ackley, David and Small, Trent},
@@ -25,7 +26,7 @@
   volume = {ALIFE 14: The Fourteenth International Conference on the Synthesis and Simulation of Living Systems},
   pages = {606--613},
   doi = {10.1162/978-0-262-32621-6-ch098},
-  url = {https://doi.org/10.1162/978-0-262-32621-6-ch098},
+  url = {https://direct.mit.edu/isal/proceedings/alife2014/606/99046},
   eprint = {https://direct.mit.edu/isal/proceedings-pdf/alife2014/26/606/1901859/978-0-262-32621-6-ch098.pdf},
 }
 @inproceedings{angluin2005power,
@@ -70,7 +71,7 @@
   volume = 19,
   number = 5,
   pages = {599--609},
-  doi = {https://doi.org/10.1016/S0167-739X(02)00171-1},
+  doi = {10.1016/s0167-739x(02)00171-1},
   issn = {0167-739x},
   url = {https://www.sciencedirect.com/science/article/pii/S0167739X02001711},
   note = {Tools for Program Development and Analysis. Best papers from two Technical Sessions, at ICCS2001, San Francisco, CA, USA, and ICCS2002, Amsterdam, The Netherlands},
@@ -98,6 +99,7 @@
   publisher = {Google Patents},
   note = {US Patent App. 12/267,287},
 }
+% dead doi
 @inproceedings{bohm2017mabe,
   title = {MABE (Modular Agent Based Evolver): A framework for digital evolution research},
   author = {Bohm, Clifford and G., Nitash C. and Hintze, Arend},
@@ -107,7 +109,7 @@
   volume = {ECAL 2017, the Fourteenth European Conference on Artificial Life},
   pages = {76--83},
   doi = {10.1162/isal_a_016},
-  url = {https://doi.org/10.1162/isal\_a\_016},
+  url = {https://direct.mit.edu/isal/proceedings/ecal2017/76/99520},
   eprint = {https://direct.mit.edu/isal/proceedings-pdf/ecal2017/29/76/1904506/isal\_a\_016.pdf},
 }
 @inproceedings{bohm2019mabe,
@@ -346,7 +348,7 @@
   booktitle = {Artificial Intelligence in the Age of Neural Networks and Brain Computing},
   publisher = {Academic Press},
   pages = {293--312},
-  doi = {https://doi.org/10.1016/B978-0-12-815480-9.00015-3},
+  doi = {10.1016/b978-0-12-815480-9.00015-3},
   isbn = {978-0-12-815480-9},
   url = {https://www.sciencedirect.com/science/article/pii/B9780128154809000153},
   editor = {Robert Kozma and Cesare Alippi and Yoonsuck Choe and Francesco Carlo Morabito},
@@ -399,6 +401,7 @@
   publisher = {Springer Berlin Heidelberg},
   address = {Berlin, Heidelberg},
   pages = {146--157},
+  doi = {10.1007/978-3-540-78671-9_13},
   isbn = {978-3-540-78671-9},
   editor = {O'Neill, Michael and Vanneschi, Leonardo and Gustafson, Steven and Esparcia Alc{\'a}zar, Anna Isabel and De Falco, Ivanoe and Della Cioppa, Antonio and Tarantino, Ernesto},
 }
