### 2022-10-18 #4860 @mmore500
All good! The two invalid doi's are directly from publisher's website --- they appear to be dead, so I have submitted error reports to the doi folks. Suggested doi is for a book review, not the source itself.
