### 2022-10-19 #4860 @GeekLogan
> 👋 @jmcbroome @jsun @tkmamidi @GeekLogan would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi 👋 @osorensen! I would be happy to review if you are still in need, however, I am not super experienced with this subject area so the suggested reviewers may be a better fit if they are available. Feel free to let me know
