### 2022-10-19 #4860 @kgd-al
> > > wave @JJ, @kgd-al, @tim-taylor would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
> >
> >
> > I did the thumbs-up here, but I guess it's not enough...
>
> Does the same apply to you @kgd-al?

Indeed. I would be happy to review although this is a first so I will probably have a few questions later on.
