### 2022-10-19 #4860 @tim-taylor
> wave @JJ, @kgd-al, @tim-taylor would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

@osorensen I'm sorry but I don't have time to do this as I'm travelling a lot over the next few weeks. But it sounds like you are sorted for reviewers now anyway 👍
