### 2022-10-24 mmore500/hstrat#45 @GeekLogan
README Figure Background
This is part of an ongoing review of this software for JOSS: https://github.com/openjournals/joss-reviews/issues/4866

Hi @mmore500, I was reading through the README and I found that some of the figures are generated without a background, leading to difficult readability if you are on any of the dark GitHub/VSCode themes (I posted an example below of what it looks like with the GitHub dark theme). would you consider adding a white or neutral background to the image files to improve this?

Thanks!

![image](https://user-images.githubusercontent.com/5133556/197552621-39a1aec1-8035-49fe-be25-6b1db295dc2f.png)
