### 2022-10-25 mmore500/hstrat#45 @mmore500
Hadn't realized that. On it! :)
