### 2022-10-25 0eaeb52
Rename assets "policy" -> "algorithm"
