### 2022-10-25 2ac0146
Add solid bg to graphics to improve readability
Closes #45
