+++ test_hstrat/test_stratum_retention_viz/test_plot/test_stratum_retention_dripplot.py
@@ -94,4 +94,5 @@ def test_docplots(policy):
             "policy": slugify(str(policy)),
             "num_generations": "256",
         },
+        teeplot_transparent=False,
     )
