+++ test_hstrat/test_stratum_retention_viz/test_plot/test_strata_retained_num_lineplot.py
@@ -93,4 +93,5 @@ def test_docplots(policy):
             "policy": slugify(str(policy)),
             "num_generations": "256",
         },
+        teeplot_transparent=False,
     )
