+++ requirements_dev.in
@@ -30,4 +30,4 @@ more-itertools==8.13.0
 pytest==7.1.2
 pytest-xdist==2.5.0
 safe-assert==0.2.0
-teeplot==0.3.0
+teeplot==0.4.0
