### 2022-10-26 mmore500/hstrat#45 @mmore500
Just tweaked the scripts and updated the assets in docs, so I think all of the images should have solid backgrounds now!
