### 2022-11-01 mmore500/hstrat#46 @GeekLogan
Comments on paper structure for JOSS
This is part of an ongoing review of this software for JOSS: https://github.com/openjournals/joss-reviews/issues/4866

Hi @mmore500, I have two small comments about the manuscript that I am just combining here for convenience. I do not believe these modifications are absolutely required, but are worth considering:

1. You might consider adding some of the overview figures in the README such as the "naive approach vs. our approach" as a Figure in the manuscript itself. I think it would help a lot with readability, especially for novice users who may not be familiar with your computational approach.

2. You list in the manuscript several projects which you have used your software for, however, it would be good to describe additional approaches/libraries which can be used for similar purposes or in conjunction with your software. For instance I found this after a short search (I will leave it up to you to decide if it is applicable): https://dl.acm.org/doi/pdf/10.1145/3377929.3398147
