### 2022-11-01 mmore500/hstrat#47 @GeekLogan
Note on installation
This is part of an ongoing review of this software for JOSS: https://github.com/openjournals/joss-reviews/issues/4866

When installing the package using `pip install hstrat` on a Mac (M1 Pro) and Ubuntu machine (22.04), I run into the following error caused by `gmpy` being compiled without the `gmp.h` header installed on my system.

>       In file included from src/gmpy.c:251:
>      ./src/gmpy.h:30:10: fatal error: 'gmp.h' file not found
>     #include "gmp.h"
>               ^~~~~~~
>      1 error generated.
>      error: command '/usr/bin/clang' failed with exit code 1
>      [end of output]

To resolve this, I installed a local `gmpy` installation using the conda package at `https://anaconda.org/conda-forge/gmpy2/`. I didn't try it, but it seems like `brew` might be another option on Mac for installing. It might be good to add a footnote to the README about this installation step and dependency on `gmp` being available in general.
