### 2022-11-01 mmore500/hstrat#47 @GeekLogan
The conda solution above did not work for Ubuntu, I needed to install the `libgmp3-dev` package to get `pip` to build the wheel properly.
