### 2022-11-01 mmore500/hstrat#48 @GeekLogan
hstrat fails to import
* hstrat version: `1.0.3`
* Python version: `Python 3.10.6`
* Operating System: `Mac OS 12.2.1` but using conda packages

### Description

I installed the package using `pip` and imported, yielding the error below. This does **NOT** occur under Python 3.9.13 using the same install procedure.

### What I Did

```
Python 3.10.6 (main, Oct  7 2022, 15:17:36) [Clang 12.0.0 ] on darwin
>>> import hstrat
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/removed/anaconda3/envs/hstrat-test/lib/python3.10/site-packages/hstrat/__init__.py", line 7, in <module>
    from . import (
  File "/Users/removed/anaconda3/envs/hstrat-test/lib/python3.10/site-packages/hstrat/_auxiliary_lib/__init__.py", line 40, in <module>
    eval(o).__module__ = __name__
TypeError: cannot set '__module__' attribute of immutable type 'itertools.pairwise'
```

This is part of an ongoing review of this software for JOSS: https://github.com/openjournals/joss-reviews/issues/4866
