### 2022-11-06 mmore500/hstrat#48 @mmore500
Noted! Definitely need to get this taken care of. Will take a crack at this and #46, #47  ASAP. Making the final push on my dissertation, so I may not be able to circle back to this until Nov 15.
