### 2022-11-08 32551e7
Will actually print something
Alternatively, it could tell the user to use an interactive shell like bpython that prints the result of an operation
