### 2022-11-19 c05448a
Catch TypeError when laundering __module__
Fixes bug with itertools.pairwise in Pythonn 3.10+. Closes #48.
