+++ setup.py
@@ -62,6 +62,6 @@ setup(
     test_suite="tests",
     tests_require=test_requirements,
     url="https://github.com/mmore500/hstrat",
-    version="1.0.3",
+    version="1.0.4",
     zip_safe=False,
 )
