### 2022-11-19 mmore500/hstrat#48 @mmore500
This should be fixed in the new version 1.0.4 which should show up on PyPi in a few minutes. Let me know if it isn't!
