### 2022-11-19 9413da2
Implement, test popcount helper
