+++ requirements_dev.in
@@ -10,6 +10,7 @@ wheel==0.33.6
 
 
 anytree==2.8.0
+bitstring==3.1.9
 iterpop==0.3.4
 interval-search==0.2.0
 gmpy==1.17
