### 2022-11-19 dddc671
Remove gmpy, replace with popcount helper
Closes #47
