+++ setup.py
@@ -15,7 +15,6 @@ requirements = [
     "bitstring>=3.1.9",
     "iterpop>=0.3.4",
     "interval_search>=0.2.0",
-    "gmpy>=1.17",
     "keyname>=0.4.1",
     "lru-dict>=1.1.7",
     "matplotlib>=3.5.2",
