### 2022-11-19 mmore500/hstrat#47 @mmore500
The easiest thing to do here was to switch out the gmpy dependency for a pure python alternative. Completely agree that the possibility for install issues makes this the better choice. Should be fixed in v1.0.5, which should show up on PyPi shortly.
