+++ joss/paper.bib
@@ -162,6 +162,20 @@
   url = {https://royalsocietypublishing.org/doi/abs/10.1098/rspb.2010.1557},
   eprint = {https://royalsocietypublishing.org/doi/pdf/10.1098/rspb.2010.1557},
 }
+@article{cock2009biopython,
+  title = {Biopython: freely available Python tools for computational molecular biology and bioinformatics},
+  author = {Cock, Peter J. A. and Antao, Tiago and Chang, Jeffrey T. and Chapman, Brad A. and Cox, Cymon J. and Dalke, Andrew and Friedberg, Iddo and Hamelryck, Thomas and Kauff, Frank and Wilczynski, Bartek and de Hoon, Michiel J. L.},
+  year = 2009,
+  month = {03},
+  journal = {Bioinformatics},
+  volume = 25,
+  number = 11,
+  pages = {1422--1423},
+  doi = {10.1093/bioinformatics/btp163},
+  issn = {1367-4803},
+  url = {https://doi.org/10.1093/bioinformatics/btp163},
+  eprint = {https://academic.oup.com/bioinformatics/article-pdf/25/11/1422/944180/btp163.pdf},
+}
 @inproceedings{de2012deap,
   title = {DEAP: A Python Framework for Evolutionary Algorithms},
   author = {De Rainville, Fran\c{c}ois-Michel and Fortin, F\'{e}lix-Antoine and Gardner, Marc-Andr\'{e} and Parizeau, Marc and Gagn\'{e}, Christian},
