### 2022-11-19 c40eafc
Add cartoon summary of hereditary stratigraphy
