### 2022-11-19 mmore500/hstrat#46 @mmore500
Agreed on both of these points. I added a cartoon figure to the introduction and a "related software section." Both of these changes are pushed to the master branch.
