### 2022-11-20 mmore500/hstrat#48 @GeekLogan
I can confirm that it works correctly on my machine under Python 3.10
