### 2022-11-21 mmore500/hstrat#52 @JJ
Tests are a bit confusing
### Description

> Related to openjournals/joss-reviews#4866

Testing does not follow standard convention.

* Tests are not placed in the usual directory, `tests`
* `pytest` can't find them, for that reason
* CI uses `tox`, but `tox` is installed on purpose, and not included in `pyproject.toml`. Running `tox` from the command line does not work
*
![Captura de pantalla de 2022-11-21 18-27-42](https://user-images.githubusercontent.com/500/203121345-5e114e67-3aba-4fc0-b27f-25c320083083.png)
* The `test` target in the root `Makefile` does not work either.
![Captura de pantalla de 2022-11-21 18-28-47](https://user-images.githubusercontent.com/500/203121526-15ce7185-ef47-4a77-9837-194fc12a6b41.png)
* None of this is mentioned in the README, or is easily accessible in a clearly named section
