### 2022-11-21 be49e06
Test with latest versions of Python
Future-proof your package
