+++ tox.ini
@@ -33,6 +33,8 @@ add_ignore = D100, D101, D103, D104, D105, D205, D400
 python =
     3.8: py38, mypy
     3.9: py39
+    3.10: py310
+    3.11: py311
 
 [testenv:flake8]
 basepython = python
