### 2022-11-21 e76507c
fixup! Move test_hstrat -> tests/test_hstrat
