+++ tox.ini
@@ -51,4 +51,4 @@ deps =
 ;     -r{toxinidir}/requirements.txt
 commands =
     pip install -U pip
-    pytest -n auto {toxinidir}/test_hstrat
+    pytest -n auto {toxinidir}/tests/test_hstrat
