### 2022-11-22 e031e7f
Migrate to pyproject.toml, compile reqs w/ py3.7
