+++ tox.ini
@@ -1,5 +1,6 @@
 [tox]
 envlist = py38, py39, flake8, mypy
+isolated_build = True
 
 [coverage:report]
 exclude_lines =
@@ -45,7 +46,7 @@ commands = flake8 hstrat tests
 setenv =
     PYTHONPATH = {toxinidir}
 deps =
-    -r{toxinidir}/requirements_dev.txt
+    -r{toxinidir}/requirements-dev/requirements-testing.txt
 ; If you want to make tox run the tests with the same versions, create a
 ; requirements.txt with the pinned versions and uncomment the following line:
 ;     -r{toxinidir}/requirements.txt
