+++ setup.cfg
@@ -15,6 +15,10 @@ replace = __version__ = "{new_version}"
 search = version: {current_version}
 replace = version: {new_version}
 
+[bumpversion:file:pyproject.toml]
+search = version = "{current_version}"
+replace = version = "{new_version}"
+
 [bdist_wheel]
 universal = 1
 
