+++ requirements_dev.in
@@ -1,30 +1,0 @@
-bump2version==0.5.11
-flake8==3.7.8
-Sphinx==4.4.0
-sphinx_rtd_theme==1.0.0
-tox==3.24.0
-twine==1.14.0
-wheel==0.33.6
-
-
-anytree==2.8.0
-bitstring==3.1.9
-iterpop==0.3.4
-interval-search==0.2.0
-keyname==0.4.1
-lru-dict==1.1.7
-matplotlib==3.5.2
-mmh3==3.0.0
-mpmath==1.1.0
-opytional==0.1.0
-python-slugify==6.1.2
-seaborn==0.11.2
-
-# testing
-coverage==6.4.4
-scipy==1.5.4
-iterify==0.1.0
-more-itertools==8.13.0
-pytest==7.1.2
-pytest-xdist==2.5.0
-teeplot==0.4.0
