### 2022-11-22 f148e39
Migrate to pyproject.toml, compile reqs w/ py3.7
