### 2022-11-22 74d877f
Test over py 3.7 thru 3.11 on CI
