+++ test_hstrat/test_stratum_retention_strategy/test_stratum_retention_algorithms/test_geom_seq_nth_root_algo/test_GenDropRanks.py
@@ -7,6 +7,9 @@ from hstrat._auxiliary_lib import all_same
 from hstrat.hstrat import geom_seq_nth_root_algo
 
 
+@pytest.mark.filterwarnings(
+    "ignore:Interspersal set to 1, no bound on MRCA rank estimate uncertainty can be guaranteed."
+)
 @pytest.mark.parametrize(
     "degree",
     [
@@ -87,6 +90,9 @@ def test_impl_consistency(degree, interspersal, time_sequence):
         )
 
 
+@pytest.mark.filterwarnings(
+    "ignore:Interspersal set to 1, no bound on MRCA rank estimate uncertainty can be guaranteed."
+)
 @pytest.mark.parametrize(
     "degree",
     [
@@ -150,6 +156,9 @@ def test_policy_consistency(degree, interspersal, time_sequence):
             ) == sorted(policy_requirement)
 
 
+@pytest.mark.filterwarnings(
+    "ignore:Interspersal set to 1, no bound on MRCA rank estimate uncertainty can be guaranteed."
+)
 @pytest.mark.parametrize(
     "degree",
     [
