+++ test_hstrat/test_stratum_retention_strategy/test_stratum_retention_algorithms/test_geom_seq_nth_root_tapered_algo/test_CalcMrcaUncertaintyAbsUpperBound.py
@@ -6,6 +6,9 @@ import pytest
 from hstrat.hstrat import geom_seq_nth_root_tapered_algo
 
 
+@pytest.mark.filterwarnings(
+    "ignore:Interspersal set to 1, no bound on MRCA rank estimate uncertainty can be guaranteed."
+)
 @pytest.mark.parametrize(
     "degree",
     [
@@ -92,6 +95,9 @@ def test_policy_consistency(degree, interspersal, time_sequence):
                 )
 
 
+@pytest.mark.filterwarnings(
+    "ignore:Interspersal set to 1, no bound on MRCA rank estimate uncertainty can be guaranteed."
+)
 @pytest.mark.parametrize(
     "degree",
     [
@@ -130,6 +136,9 @@ def test_eq(degree, interspersal):
     assert instance is not None
 
 
+@pytest.mark.filterwarnings(
+    "ignore:Interspersal set to 1, no bound on MRCA rank estimate uncertainty can be guaranteed."
+)
 @pytest.mark.parametrize(
     "degree",
     [
