+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py38, py39, flake8, mypy
+envlist = py37, py38, py39, flake8, mypy, py310, py311
 
 [coverage:report]
 exclude_lines =
