+++ .github/workflows/CI.yml
@@ -21,6 +21,16 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install tox tox-gh-actions
+    # numpy build dependencies
+    - run: |
+        sudo apt-get install \
+          build-essential \
+          fort77 \
+          gcc \
+          libatlas-base-dev \
+          libblas-dev \
+          liblapack-dev
+        python3-dev
     - name: Test with tox
       run: tox
 
