+++ .github/workflows/CI.yml
@@ -22,7 +22,9 @@ jobs:
         python -m pip install --upgrade pip
         pip install tox tox-gh-actions
     # numpy build dependencies
+    # xlocal workaround adapted from https://github.com/carla-simulator/carla/issues/503#issuecomment-397658790
     - run: |
+        ln -s /usr/include/locale.h /usr/include/xlocale.h
         sudo apt-get install \
           build-essential \
           fort77 \
