+++ tox.ini
@@ -42,14 +42,29 @@ basepython = python
 deps = flake8
 commands = flake8 hstrat tests
 
+[testenv:py37]
+deps =
+    -r{toxinidir}/requirements-dev/py37/requirements-testing.txt
+
+[testenv:py38]
+deps =
+    -r{toxinidir}/requirements-dev/py38/requirements-testing.txt
+
+[testenv:py39]
+deps =
+    -r{toxinidir}/requirements-dev/py39/requirements-testing.txt
+
+[testenv:py310]
+deps =
+    -r{toxinidir}/requirements-dev/py310/requirements-testing.txt
+
+[testenv:py311]
+deps =
+    -r{toxinidir}/requirements-dev/py311/requirements-testing.txt
+
 [testenv]
 setenv =
     PYTHONPATH = {toxinidir}
-deps =
-    -r{toxinidir}/requirements-dev/requirements-testing.txt
-; If you want to make tox run the tests with the same versions, create a
-; requirements.txt with the pinned versions and uncomment the following line:
-;     -r{toxinidir}/requirements.txt
 commands =
     pip install -U pip
     pytest -n auto {toxinidir}/tests/test_hstrat
