### 2022-11-22 bd47928
Revert numpy build dependencies in gh actions
Revert "Use sudo for xlocale workaround"

This reverts commit 7e1d61ebe16703b1873920bcd1798f32d90e972b.

Revert "Add xlocale workaround"

This reverts commit aac84f8f1ebf1dd854ccc5943590b15c9ec89928.

Revert "Add gfortran dependency"

This reverts commit 78ef5be13bf163d7d74949728852645547bf6e6b.

Revert "Fix syntax"

This reverts commit cf8b5e922e336d3a9ce541fd38df7196c7a7d4db.

Revert "Install numpy build dependencies"

This reverts commit aa7ee76618feb82022cac7b55f435f40e16ad9a3.
