+++ .github/workflows/CI.yml
@@ -21,19 +21,6 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install tox tox-gh-actions
-    # numpy build dependencies
-    # xlocal workaround adapted from https://github.com/carla-simulator/carla/issues/503#issuecomment-397658790
-    - run: |
-        sudo ln -s /usr/include/locale.h /usr/include/xlocale.h
-        sudo apt-get install \
-          build-essential \
-          fort77 \
-          gcc \
-          gfortran \
-          libatlas-base-dev \
-          libblas-dev \
-          liblapack-dev \
-          python3-dev
     - name: Test with tox
       run: tox
 
