+++ tox.ini
@@ -32,6 +32,7 @@ add_ignore = D100, D101, D103, D104, D105, D205, D400
 
 [gh-actions]
 python =
+    3.7: py37
     3.8: py38, mypy
     3.9: py39
     3.10: py310
