### 2022-11-22 f694134
fixup! Install compatible deps for each py version
