+++ pyproject.toml
@@ -144,6 +144,8 @@ docs = [
 
 [tool.black]
 line_length = 79
+# due to some weird circular symlink issue
+extend_exclude = "requirements-dev"
 
 [tool.isort]
 force_sort_within_sections = true
