### 2022-11-22 e398447
Fix mpl too many open figures warning
