### 2022-11-22 cf7f13c
Actually silence mpl discarded animation warning
Warning popping up elsewhere due to lazy garbage collection
