+++ tests/test_hstrat/test_stratum_retention_viz/test_animate/test_policy_panel_animate.py
@@ -3,9 +3,6 @@ import pytest
 from hstrat import hstrat
 
 
-@pytest.mark.filterwarnings(
-    "ignore:Animation was deleted without rendering anything."
-)
 @pytest.mark.parametrize(
     "policy",
     [
@@ -13,10 +10,9 @@ from hstrat import hstrat
     ],
 )
 def test_one(policy):
-    hstrat.policy_panel_animate(
-        policy,
-        10,
-    )
+    hstrat.policy_panel_animate(policy, 10,).to_html5_video(
+        embed_limit=0
+    )  # silence mpl unused animation warning
 
 
 @pytest.mark.heavy
