### 2022-11-22 ab5e04d
Actually silence mpl discarded animation warning
Warning popping up elsewhere due to lazy garbage collection
