+++ tests/test_hstrat/test_stratum_retention_viz/test_animate/test_stratum_retention_animate.py
@@ -3,9 +3,6 @@ import pytest
 from hstrat import hstrat
 
 
-@pytest.mark.filterwarnings(
-    "ignore:Animation was deleted without rendering anything."
-)
 @pytest.mark.parametrize(
     "policy",
     [
@@ -19,18 +16,21 @@ def test_one(policy):
         draw_extant_history=False,
         draw_extinct_history=False,
         draw_extinct_placeholders=True,
-    )
+    ).to_html5_video(
+        embed_limit=0
+    )  # silence mpl unused animation warning
     hstrat.stratum_retention_animate(
         policy,
         10,
         draw_extant_history=True,
         draw_extinct_history=False,
         draw_extinct_placeholders=True,
-    )
-    hstrat.stratum_retention_animate(
-        policy,
-        10,
-    )
+    ).to_html5_video(
+        embed_limit=0
+    )  # silence mpl unused animation warning
+    hstrat.stratum_retention_animate(policy, 10).to_html5_video(
+        embed_limit=0
+    )  # silence mpl unused animation warning
 
 
 @pytest.mark.heavy
