+++ tests/test_hstrat/test_stratum_retention_viz/test_plot/test_mrca_uncertainty_absolute_barplot.py
@@ -14,6 +14,7 @@ def _release_cur_mpl_fig():
     plt.clf()
     plt.close()
 
+
 @pytest.mark.parametrize(
     "policy",
     [
