### 2022-11-22 7a88df9
Try passing string as seed instead of hashing
