+++ tests/test_hstrat/conftest.py
@@ -9,6 +9,5 @@ import pytest
 @pytest.fixture(autouse=True)
 def reseed_random(request: pytest.FixtureRequest) -> None:
     """Reseed random number generator to ensure determinstic test."""
-    seed = hash(request.node.name) % 2**32
-    random.seed(seed)
-    np.random.seed(seed)
+    random.seed(request.node.name)
+    np.random.seed(random.randint(1, 2**32))
