### 2022-11-22 9e91716
Disable fail-fast over python version matrix
