+++ .github/workflows/CI.yml
@@ -8,6 +8,7 @@ jobs:
   test:
     runs-on: ubuntu-latest
     strategy:
+      fail-fast: true
       matrix:
         python-version: [3.7, 3.8, 3.9, "3.10", 3.11]
 
