+++ .github/workflows/CI.yml
@@ -19,6 +19,7 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
+        sudo apt-get install -y ffmpeg
         python -m pip install --upgrade pip
         pip install tox tox-gh-actions
     - name: Test with tox
@@ -34,6 +35,7 @@ jobs:
         python-version: 3.9
     - name: Install dependencies
       run: |
+        sudo apt-get install -y ffmpeg
         python -m pip install --upgrade pip
         python -m pip install -r requirements-dev/requirements-testing.txt
     - name: Log coverage configuration
