### 2022-11-22 7b45557
Update CONTRIBUTING.rst
