### 2022-11-22 3949f10
Strip trailing whitespace
