### 2022-11-22 b4fc935
Add requirements for mypy
