+++ tox.ini
@@ -51,6 +51,10 @@ deps =
 deps =
     -r{toxinidir}/requirements-dev/py38/requirements-testing.txt
 
+[testenv:mypy]
+deps =
+    -r{toxinidir}/requirements-dev/py38/requirements-testing.txt
+
 [testenv:py39]
 deps =
     -r{toxinidir}/requirements-dev/py39/requirements-testing.txt
