+++ tox.ini
@@ -71,5 +71,4 @@ deps =
 setenv =
     PYTHONPATH = {toxinidir}
 commands =
-    pip install -U pip
     pytest -n auto {toxinidir}/tests/test_hstrat
