+++ pyproject.toml
@@ -86,7 +86,7 @@ pinned_dependencies_py37 = [
   # however this pin is necessary due an apparent bug in pip-compile
   "pandas==1.3.5",
   "python-slugify==6.1.2",
-  "seaborn==0.11.2",
+  "seaborn==0.12.1",
 ]
 pinned_dependencies_py310 = [
   "anytree==2.8.0",
