### 2022-11-22 5b833a6
Add mypy to testing requirements
