+++ pyproject.toml
@@ -111,6 +111,7 @@ testing_py37 = [
   "flake8==3.7.8",
   "iterify==0.1.0",
   "more-itertools==8.13.0",
+  "mypy==0.991",
   "pytest-xdist==2.5.0",
   "pytest==7.1.2",
   "safe-assert==0.2.0",
@@ -123,6 +124,7 @@ testing_py310 = [
   "flake8==3.7.8",
   "iterify==0.1.0",
   "more-itertools==8.13.0",
+  "mypy==0.991",
   "pytest-xdist==2.5.0",
   "pytest==7.1.2",
   "safe-assert==0.2.0",
