### 2022-11-22 #4866 @mmore500
Yup! @jj had some super useful suggestions,  sat down to incorporate them this evening, think they should all be worked in!
