### 2022-11-23 mmore500/hstrat#61 @kgd-al
Examples 
Part of the ongoing joss review (https://github.com/openjournals/joss-reviews/issues/4866)

I have a few remarks stemming from the provided example so I address them all here:

#### 1. example/basics.py
The example provided in the description could be a standalone file (in a dedicated folder) to facilitate forking.

#### 2. Non interactive print
In the same line as #50, it would be better for the common ancestry tests to use `print` as well:
```
print(hstrat.does_have_any_common_ancestor(individual1, individual2)) # -> False
print(hstrat.does_have_any_common_ancestor(individual1_child1, individual2)) # -> False
```

#### 3. True case(s?)
Adding a `True` case would also help with understanding
```
print(hstrat.does_have_any_common_ancestor(individual1_child1, individual1)) # -> True
```

#### 4.
What are your thoughts on a human-friendly way to see the contents of the stratum (.e.g `__repr__` or a `ToString`)?

#### 5.
May be provide a practical example:
```
class Genome:
  def __init__(self):
    self._data = 0
    self._stratum = hstrat...

 def mutate(self):
   self._data += 1
   self._stratum ???

 ....

a = Genome()
b = Genome()
for i in range(100):
  a.mutate()
  b.mutate()

# do something fascinating (may be with a graph?)
```

So that users can see at a glance how to include your library in their code and how to make something of it fast.
