### 2022-11-24 79cdcbc
Fixup CONTRIBUTING.rst
