### 2022-11-24 c07e688
Add option to install test reqs into local env
