### 2022-11-24 4b89dd7
Add option to install test reqs into local env
