### 2022-11-24 3b06c71
Warn if testing requirements are missing
