+++ pyproject.toml
@@ -115,6 +115,7 @@ testing_py37 = [
   "pytest-xdist==2.5.0",
   "pytest==7.1.2",
   "safe-assert==0.2.0",
+  "setuptools",
   "scipy==1.7.3",
   "teeplot==0.4.0",
   "tox==3.24.0",
@@ -128,6 +129,7 @@ testing_py310 = [
   "pytest-xdist==2.5.0",
   "pytest==7.1.2",
   "safe-assert==0.2.0",
+  "setuptools",
   "scipy==1.9.3",
   "teeplot==0.4.0",
   "tox==3.24.0",
