### 2022-11-24 fa81695
Warn if testing requirements are missing
