### 2022-11-24 mmore500/hstrat#60 @mmore500
Agreed that this should be improved. Put together a PR (#62) that will detect missing testing requirements and direct the user to documentation on how to install them.
