### 2022-11-24 c99ecd0
Release mpl figs to fix open figures warning
