### 2022-11-24 e4b7c6b
Release mpl figs to fix open figures warning
