+++ tests/test_hstrat/test_stratum_retention_viz/test_plot/test_policy_panel_plot.py
@@ -1,6 +1,7 @@
 import unittest
 
 from hstrat import hstrat
+from hstrat._auxiliary_lib import release_cur_mpl_fig
 
 
 class TestPolicyPanelPlot(unittest.TestCase):
@@ -15,7 +16,9 @@ class TestPolicyPanelPlot(unittest.TestCase):
             hstrat.perfect_resolution_algo.Policy(),
         ]:
             hstrat.policy_panel_plot(policy, 100, do_show=False)
+            release_cur_mpl_fig()
             hstrat.policy_panel_plot(policy, 10, do_show=False)
+            release_cur_mpl_fig()
 
 
 if __name__ == "__main__":
