### 2022-11-24 mmore500/hstrat#60 @mmore500
#63 should fix that annoying matplotlib warning!
