### 2022-11-24 mmore500/hstrat#61 @mmore500
I like the idea of putting examples in an examples folder. This will also make it easier to make sure they stay up to date over time (because we can run the examples as part of CI). Working on this now!
