### 2022-11-25 79fc68c
Implement, test does_def_share_no_common_ancestry
