### 2022-11-25 4748a22
Implement, test does_def_share_no_common_ancestry
