+++ tests/test_hstrat/test_phylogenetic_inference/test_population/test_calc_rank_of_earliest_detectable_mrca_among.py
@@ -1,0 +1,186 @@
+import itertools as it
+import random
+
+import pytest
+
+from hstrat import hstrat
+
+
+@pytest.mark.filterwarnings(
+    "ignore:Insufficient common ranks between columns to detect common ancestry at given confidence level."
+)
+@pytest.mark.parametrize(
+    "confidence_level",
+    [0.8, 0.95, 0.99],
+)
+@pytest.mark.parametrize(
+    "differentia_bit_width",
+    [1, 2, 8, 64],
+)
+def test_calc_rank_of_earliest_detectable_mrca_among1(
+    confidence_level,
+    differentia_bit_width,
+):
+    expected_thresh = (
+        hstrat.HereditaryStratigraphicColumn(
+            stratum_differentia_bit_width=differentia_bit_width,
+        ).CalcMinImplausibleSpuriousConsecutiveDifferentiaCollisions(
+            significance_level=1 - confidence_level,
+        )
+        - 1
+    )
+
+    for r1, r2, r3 in it.product(*[range(1, expected_thresh)] * 3):
+        c1 = hstrat.HereditaryStratigraphicColumn(
+            stratum_differentia_bit_width=differentia_bit_width,
+            stratum_retention_policy=hstrat.perfect_resolution_algo.Policy(),
+        )
+        c2 = c1.Clone()
+        c3 = hstrat.HereditaryStratigraphicColumn(
+            stratum_differentia_bit_width=differentia_bit_width,
+            stratum_retention_policy=hstrat.fixed_resolution_algo.Policy(2),
+        )
+        for __ in range(r1 - 1):
+            c1.DepositStratum()
+        for __ in range(r2 - 1):
+            c2.DepositStratum()
+        for __ in range(r3 - 1):
+            c3.DepositStratum()
+
+        for x1, x2 in it.combinations([c1, c2, c3], 2):
+            assert (
+                hstrat.calc_rank_of_earliest_detectable_mrca_among(
+                    [x1, x2],
+                    confidence_level=confidence_level,
+                )
+                is None
+            )
+            assert (
+                hstrat.calc_rank_of_earliest_detectable_mrca_among(
+                    [x2, x1],
+                    confidence_level=confidence_level,
+                )
+                is None
+            )
+            assert (
+                hstrat.calc_rank_of_earliest_detectable_mrca_among(
+                    [x1, x2, x1],
+                    confidence_level=confidence_level,
+                )
+                is None
+            )
+            assert (
+                hstrat.calc_rank_of_earliest_detectable_mrca_among(
+                    [x2, x1] * 10,
+                    confidence_level=confidence_level,
+                )
+                is None
+            )
+
+
+@pytest.mark.parametrize(
+    "confidence_level",
+    [0.8, 0.95, 0.99],
+)
+@pytest.mark.parametrize(
+    "differentia_bit_width",
+    [1, 2, 8, 64],
+)
+def test_calc_rank_of_earliest_detectable_mrca_among2(
+    confidence_level,
+    differentia_bit_width,
+):
+    expected_thresh = (
+        hstrat.HereditaryStratigraphicColumn(
+            stratum_differentia_bit_width=differentia_bit_width,
+        ).CalcMinImplausibleSpuriousConsecutiveDifferentiaCollisions(
+            significance_level=1 - confidence_level,
+        )
+        - 1
+    )
+
+    c1 = hstrat.HereditaryStratigraphicColumn(
+        stratum_differentia_bit_width=differentia_bit_width,
+        stratum_retention_policy=hstrat.perfect_resolution_algo.Policy(),
[…]
