+++ hstrat/phylogenetic_inference/population/_calc_rank_of_earliest_detectable_mrca_among.py
@@ -1,0 +1,65 @@
+import itertools as it
+import typing
+import warnings
+
+from ...genome_instrumentation import HereditaryStratigraphicColumn
+from ...juxtaposition import get_nth_common_rank_between
+from ..pairwise import calc_rank_of_earliest_detectable_mrca_between
+
+
+def calc_rank_of_earliest_detectable_mrca_among(
+    population: typing.Iterable[HereditaryStratigraphicColumn],
+    confidence_level: float = 0.95,
+) -> typing.Optional[int]:
+    """After what generation is common ancstry robustly detectable?
+
+    Calculates the earliest possible rank a MRCA between first and second
+    could be reliably detected at.
+
+    Even if a true MRCA of first and second exists, if it occured earlier
+    than the rank calculated here it could not be reliably detected with
+    sufficient confidence after accounting for the possibility of spurious
+    differentia collisions. (Although subsequent spurious differentia
+    collisions after the true MRCA of first and second could lead to MRCA
+    detection at such a rank.)
+
+    Returns None if insufficient common ranks exist between first and second
+    to ever conclude at the given confidence level the existance of any
+    common ancestry between first and second (even if all strata at common
+    ranks had equivalent differentiae).
+
+    calc_rank_of_earliest_detectable_mrca_between :
+        Could any MRCA be detected between first and second? What is the rank
+        of the earliest MRCA that could be reliably detected?
+
+    """
+
+    earliest_detectable_mrca_rank = -1
+    any_combinations = False
+    for (first, second) in it.combinations(population, 2):
+        any_combinations = True
+        assert (
+            first.GetStratumDifferentiaBitWidth()
+            == second.GetStratumDifferentiaBitWidth()
+        )
+        earliest_detectable_mrca_rank_between = (
+            calc_rank_of_earliest_detectable_mrca_between(
+                first, second, confidence_level
+            )
+        )
+        if earliest_detectable_mrca_rank_between is None:
+            return None
+        else:
+            earliest_detectable_mrca_rank = max(
+                earliest_detectable_mrca_rank,
+                earliest_detectable_mrca_rank_between,
+            )
+
+    if not any_combinations:
+        warnings.warn(
+            "Empty or singleton population. Unable to calculate rank of "
+            "earliest detectable MRCA."
+        )
+        return None
+
+    return earliest_detectable_mrca_rank
