### 2022-11-25 a0e26b8
Fix docstring typo
