### 2022-11-25 ccbd6f2
Implement, test calc_rank_of_mrca_bounds_among
