+++ tests/test_hstrat/test_phylogenetic_inference/test_population/test_calc_rank_of_mrca_bounds_among.py
@@ -1,0 +1,637 @@
+import itertools as it
+import random
+
+from iterify import cyclify, iterify
+import opytional as opyt
+import pytest
+from scipy import stats
+
+from hstrat import hstrat
+
+
+@pytest.mark.filterwarnings(
+    "ignore:Insufficient common ranks between columns to detect common ancestry at given confidence level."
+)
+@pytest.mark.parametrize(
+    "retention_policy",
+    [
+        hstrat.perfect_resolution_algo.Policy(),
+        hstrat.nominal_resolution_algo.Policy(),
+        hstrat.fixed_resolution_algo.Policy(fixed_resolution=10),
+    ],
+)
+@pytest.mark.parametrize(
+    "ordered_store",
+    [
+        hstrat.HereditaryStratumOrderedStoreDict,
+        hstrat.HereditaryStratumOrderedStoreList,
+        pytest.param(
+            hstrat.HereditaryStratumOrderedStoreTree,
+            marks=pytest.mark.heavy,
+        ),
+    ],
+)
+def test_comparison_commutativity_asyncrhonous(
+    retention_policy,
+    ordered_store,
+):
+    population = [
+        hstrat.HereditaryStratigraphicColumn(
+            stratum_ordered_store_factory=ordered_store,
+            stratum_retention_policy=retention_policy,
+        )
+        for __ in range(10)
+    ]
+
+    for generation in range(100):
+        for first, second in it.combinations(population, 2):
+            # assert commutativity
+            assert (
+                hstrat.calc_rank_of_mrca_bounds_among([first, second])
+                == hstrat.calc_rank_of_mrca_bounds_among([second, first])
+                == hstrat.calc_rank_of_mrca_bounds_among([second, first] * 3)
+            )
+
+        # advance generation
+        random.shuffle(population)
+        for target in range(5):
+            population[target] = population[-1].CloneDescendant()
+        for individual in population:
+            # asynchronous generations
+            if random.choice([True, False]):
+                individual.DepositStratum()
+
+
+@pytest.mark.filterwarnings(
+    "ignore:Insufficient common ranks between columns to detect common ancestry at given confidence level."
+)
+@pytest.mark.parametrize(
+    "retention_policy",
+    [
+        hstrat.perfect_resolution_algo.Policy(),
+        hstrat.nominal_resolution_algo.Policy(),
+        hstrat.fixed_resolution_algo.Policy(fixed_resolution=10),
+    ],
+)
+@pytest.mark.parametrize(
+    "ordered_store",
+    [
+        hstrat.HereditaryStratumOrderedStoreDict,
+        hstrat.HereditaryStratumOrderedStoreList,
+        pytest.param(
+            hstrat.HereditaryStratumOrderedStoreTree,
+            marks=pytest.mark.heavy,
+        ),
+    ],
+)
+def test_CalcRankOfMrcaBoundsWith(retention_policy, ordered_store):
+    def make_bundle():
+        return hstrat.HereditaryStratigraphicColumnBundle(
+            {
+                "test": hstrat.HereditaryStratigraphicColumn(
+                    initial_stratum_annotation=0,
+                    stratum_ordered_store_factory=ordered_store,
+                    stratum_retention_policy=retention_policy,
+                ),
+                "control": hstrat.HereditaryStratigraphicColumn(
+                    initial_stratum_annotation=0,
+                    stratum_ordered_store_factory=ordered_store,
+                    stratum_retention_policy=hstrat.perfect_resolution_algo.Policy(),
+                ),
+            }
+        )
+
+    column = make_bundle()
+    frozen_copy = column.Clone()
+    frozen_unrelated = make_bundle()
+    population = [column.Clone() for __ in range(10)]
+    forked_isolated = column.Clone()
+    unrelated_isolated = make_bundle()
+
[…]
