### 2022-11-25 67a2d80
Implement, test does_share_any_common_ancestor
