+++ hstrat/phylogenetic_inference/population/__init__.py
@@ -8,11 +8,13 @@ from ._calc_rank_of_mrca_bounds_among import calc_rank_of_mrca_bounds_among
 from ._does_definitively_share_no_common_ancestor import (
     does_definitively_share_no_common_ancestor,
 )
+from ._does_share_any_common_ancestor import does_share_any_common_ancestor
 
 __all__ = [
     "calc_rank_of_earliest_detectable_mrca_among",
     "calc_rank_of_mrca_bounds_among",
     "does_definitively_share_no_common_ancestor",
+    "does_share_any_common_ancestor",
 ]
 
 from ..._auxiliary_lib import launder_impl_modules as _launder
