+++ tests/test_hstrat/test_phylogenetic_inference/test_population/test_does_share_any_common_ancestor.py
@@ -1,0 +1,191 @@
+import pytest
+
+from hstrat import hstrat
+
+
+@pytest.mark.parametrize(
+    "retention_policy",
+    [
+        hstrat.perfect_resolution_algo.Policy(),
+        hstrat.nominal_resolution_algo.Policy(),
+        hstrat.fixed_resolution_algo.Policy(fixed_resolution=10),
+    ],
+)
+@pytest.mark.parametrize(
+    "ordered_store",
+    [
+        hstrat.HereditaryStratumOrderedStoreDict,
+        hstrat.HereditaryStratumOrderedStoreList,
+        hstrat.HereditaryStratumOrderedStoreTree,
+    ],
+)
+def test_does_share_any_common_ancestor(retention_policy, ordered_store):
+    first = hstrat.HereditaryStratigraphicColumn(
+        stratum_ordered_store_factory=ordered_store,
+        stratum_retention_policy=retention_policy,
+    )
+    first_copy = first.Clone()
+    second = hstrat.HereditaryStratigraphicColumn(
+        stratum_ordered_store_factory=ordered_store,
+        stratum_retention_policy=retention_policy,
+    )
+
+    def assert_validity():
+        assert hstrat.does_share_any_common_ancestor([first, first, first])
+        assert hstrat.does_share_any_common_ancestor(
+            [first_copy, first_copy, first_copy]
+        )
+        assert hstrat.does_share_any_common_ancestor(
+            [first, first_copy, first]
+        )
+        assert hstrat.does_share_any_common_ancestor(
+            [first_copy, first, first]
+        )
+        assert hstrat.does_share_any_common_ancestor(
+            [first_copy, first, first] * 3
+        )
+
+        assert not hstrat.does_share_any_common_ancestor(
+            [second, first, first]
+        )
+        assert not hstrat.does_share_any_common_ancestor(
+            [first, second, first]
+        )
+        assert not hstrat.does_share_any_common_ancestor(
+            [second, first_copy, first]
+        )
+        assert not hstrat.does_share_any_common_ancestor(
+            [first_copy, second, first]
+        )
+        assert not hstrat.does_share_any_common_ancestor(
+            [second, first, second]
+        )
+        assert not hstrat.does_share_any_common_ancestor(
+            [first, second, second]
+        )
+        assert not hstrat.does_share_any_common_ancestor(
+            [second, first_copy, second]
+        )
+        assert not hstrat.does_share_any_common_ancestor(
+            [first_copy, second, second] * 3
+        )
+
+        assert hstrat.does_share_any_common_ancestor([first, first])
+        assert hstrat.does_share_any_common_ancestor([first_copy, first_copy])
+        assert hstrat.does_share_any_common_ancestor([first, first_copy])
+        assert hstrat.does_share_any_common_ancestor([first_copy, first])
+
+        assert not hstrat.does_share_any_common_ancestor([second, first])
+        assert not hstrat.does_share_any_common_ancestor([first, second])
+        assert not hstrat.does_share_any_common_ancestor([second, first_copy])
+        assert not hstrat.does_share_any_common_ancestor([first_copy, second])
+
+    assert_validity()
+    first.DepositStratum()
+    assert_validity()
+    second.DepositStratum()
+    assert_validity()
+    first.DepositStratum()
+    assert_validity()
+
+    for __ in range(100):
[…]
