+++ tests/test_hstrat/test_phylogenetic_inference/test_population/test_calc_rank_of_mrca_uncertainty_among.py
@@ -1,0 +1,605 @@
+import itertools as it
+import random
+
+import pytest
+
+from hstrat import hstrat
+
+
+@pytest.mark.parametrize(
+    "retention_policy",
+    [
+        hstrat.perfect_resolution_algo.Policy(),
+        hstrat.nominal_resolution_algo.Policy(),
+        hstrat.fixed_resolution_algo.Policy(fixed_resolution=10),
+    ],
+)
+@pytest.mark.parametrize(
+    "ordered_store",
+    [
+        hstrat.HereditaryStratumOrderedStoreDict,
+        hstrat.HereditaryStratumOrderedStoreList,
+        pytest.param(
+            hstrat.HereditaryStratumOrderedStoreTree,
+            marks=pytest.mark.heavy,
+        ),
+    ],
+)
+def test_comparison_commutativity_asyncrhonous(
+    retention_policy,
+    ordered_store,
+):
+    population = [
+        hstrat.HereditaryStratigraphicColumn(
+            stratum_ordered_store_factory=ordered_store,
+            stratum_retention_policy=retention_policy,
+        )
+        for __ in range(10)
+    ]
+
+    for generation in range(100):
+        for first, second in it.combinations(population, 2):
+            # assert commutativity
+            assert (
+                hstrat.calc_rank_of_mrca_uncertainty_among([first, second])
+                == hstrat.calc_rank_of_mrca_uncertainty_among([second, first])
+                == hstrat.calc_rank_of_mrca_uncertainty_among(
+                    [second, first, first]
+                )
+                == hstrat.calc_rank_of_mrca_uncertainty_among(
+                    [second, first] * 3
+                )
+            )
+
+        # advance generation
+        random.shuffle(population)
+        for target in range(5):
+            population[target] = population[-1].CloneDescendant()
+        for individual in population:
+            # asynchronous generations
+            if random.choice([True, False]):
+                individual.DepositStratum()
+
+
+@pytest.mark.parametrize(
+    "retention_policy",
+    [
+        pytest.param(
+            hstrat.perfect_resolution_algo.Policy(),
+            marks=pytest.mark.heavy_2a,
+        ),
+        hstrat.nominal_resolution_algo.Policy(),
+        hstrat.fixed_resolution_algo.Policy(fixed_resolution=10),
+    ],
+)
+@pytest.mark.parametrize(
+    "ordered_store",
+    [
+        pytest.param(
+            hstrat.HereditaryStratumOrderedStoreDict,
+            marks=pytest.mark.heavy_2b,
+        ),
+        hstrat.HereditaryStratumOrderedStoreList,
+        pytest.param(
+            hstrat.HereditaryStratumOrderedStoreTree,
+            marks=pytest.mark.heavy,
+        ),
+    ],
+)
+def test_comparison_commutativity_syncrhonous(
+    retention_policy,
+    ordered_store,
+):
+
+    population = [
+        hstrat.HereditaryStratigraphicColumn(
+            stratum_ordered_store_factory=ordered_store,
+            stratum_retention_policy=retention_policy,
+        )
+        for __ in range(10)
+    ]
+
+    for generation in range(100):
+
+        for first, second in it.combinations(population, 2):
+            # assert commutativity
+            assert (
+                hstrat.calc_rank_of_mrca_uncertainty_among([first, second])
+                == hstrat.calc_rank_of_mrca_uncertainty_among([second, first])
+                == hstrat.calc_rank_of_mrca_uncertainty_among(
+                    [second, first, second]
+                )
+                == hstrat.calc_rank_of_mrca_uncertainty_among(
+                    [second, first] * 3
+                )
+            )
+
+        # advance generation
+        random.shuffle(population)
[…]
