+++ hstrat/phylogenetic_inference/__init__.py
@@ -3,5 +3,6 @@
 from . import pairwise, population
 from .pairwise import *  # noqa: F401
 from .population import *  # noqa: F401
+from .tree_reconstruction import *  # noqa: F401
 
-__all__ = pairwise.__all__ + population.__all__
+__all__ = pairwise.__all__ + population.__all__ + tree_reconstruction.__all__
