+++ hstrat/phylogenetic_inference/tree_reconstruction/__init__.py
@@ -1,0 +1,8 @@
+"""Functions to reconstruct a phylogenetic tree from extant hereditary strata."""
+
+__all__ = []
+
+from ..._auxiliary_lib import launder_impl_modules as _launder
+
+_launder([eval(item) for item in __all__], __name__)
+del _launder  # prevent name from leaking
