### 2022-11-25 84022bd
Bugfix: use tee to handle generators correctly
