+++ tests/test_hstrat/test_phylogenetic_inference/test_population/test_does_definitively_share_no_common_ancestor.py
@@ -235,3 +235,27 @@ def test_does_definitively_share_no_common_ancestor4(retention_policy):
         c2 = c1.Clone()
         assert hstrat.does_definitively_share_no_common_ancestor([]) is None
         assert hstrat.does_definitively_share_no_common_ancestor([]) is None
+
+
+def test_does_definitively_share_no_common_ancestor_generator():
+    c1 = hstrat.HereditaryStratigraphicColumn(
+        stratum_differentia_bit_width=1,
+    )
+    for __ in range(10):
+        assert hstrat.does_definitively_share_no_common_ancestor(
+            [c1 for __ in range(10)]
+        ) == hstrat.does_definitively_share_no_common_ancestor(
+            (c1 for __ in range(10))
+        )
+        c1.DepositStratum()
+
+    c2 = hstrat.HereditaryStratigraphicColumn(
+        stratum_differentia_bit_width=64,
+    )
+    for __ in range(10):
+        assert hstrat.does_definitively_share_no_common_ancestor(
+            [c2 for __ in range(10)]
+        ) == hstrat.does_definitively_share_no_common_ancestor(
+            (c2 for __ in range(10))
+        )
+        c2.DepositStratum()
