### 2022-11-25 617be63
Bugfix: use tee to handle generators correctly
