+++ tests/test_hstrat/test_phylogenetic_inference/test_population/test_calc_rank_of_mrca_uncertainty_among.py
@@ -597,9 +597,33 @@ def test_calc_rank_of_mrca_uncertainty_among_singleton():
     )
     for __ in range(10):
         assert hstrat.calc_rank_of_mrca_uncertainty_among([c1]) is None
-        c1.DepositStratum()
+        c2.DepositStratum()
 
 
 @pytest.mark.filterwarnings("ignore:Empty or singleton population.")
 def test_calc_rank_of_mrca_uncertainty_among_empty():
     assert hstrat.does_share_any_common_ancestor([]) is None
+
+
+def test_calc_rank_of_mrca_uncertainty_among_generator():
+    c1 = hstrat.HereditaryStratigraphicColumn(
+        stratum_differentia_bit_width=1,
+    )
+    for __ in range(10):
+        assert hstrat.calc_rank_of_mrca_uncertainty_among(
+            [c1 for __ in range(10)]
+        ) == hstrat.calc_rank_of_mrca_uncertainty_among(
+            (c1 for __ in range(10))
+        )
+        c1.DepositStratum()
+
+    c2 = hstrat.HereditaryStratigraphicColumn(
+        stratum_differentia_bit_width=64,
+    )
+    for __ in range(10):
+        assert hstrat.calc_rank_of_mrca_uncertainty_among(
+            [c2 for __ in range(10)]
+        ) == hstrat.calc_rank_of_mrca_uncertainty_among(
+            (c2 for __ in range(10))
+        )
+        c2.DepositStratum()
