### 2022-11-25 9acad7f
Simplify README example, create more in examples/
Also, test examples in CI
