### 2022-11-25 6369ec1
Simplify README example, create more in examples/
Also, test examples in CI
