+++ pyproject.toml
@@ -107,6 +107,7 @@ pinned_dependencies_py310 = [
   "seaborn==0.12.1",
 ]
 testing_py37 = [
+  "biopython==1.79",
   "coverage==6.4.4",
   "flake8==3.7.8",
   "iterify==0.1.0",
@@ -121,6 +122,7 @@ testing_py37 = [
   "tox==3.24.0",
 ]
 testing_py310 = [
+  "biopython==1.79",
   "coverage==6.4.4",
   "flake8==3.7.8",
   "iterify==0.1.0",
