### 2022-11-25 mmore500/hstrat#61 @mmore500
Just merged #64 into the master branch. It updates the example in the README to be a little simpler and have True cases. The bigger change is creating a new `examples/` folder with some of the more practical examples you suggested.

Let me know your thoughts on the changes to the examples and if you have any ideas for further improvements.

I think adding a nice way to print out a column and a stratum could be useful and make for a good additional example where it just shows the state of a column over a few generations step-by-step. Will update you when I get that together.

Thank you for the suggestions! I think this will be a big improvement to the usability of the library.
