### 2022-11-25 aa097ed
Fixup typo in example
