### 2022-11-29 mmore500/hstrat#61 @kgd-al
I have reviewed #64 and it addresses all of my above remarks, except 4.
What are your thoughts on that? If I want a quick and dirty look at the column or serialize it to my custom format, what is the entry point?
