### 2022-12-06 #4866 @kgd-al
@osorensen @mmore500 While I am still waiting for an answer to one last piece of details in https://github.com/mmore500/hstrat/issues/61, it is clearly a non blocking issue.

Thus, as everything looks good, I consider my review completed.
