### 2022-12-06 84d0430
Implement, test IterRetainedStrata
