### 2022-12-06 b5b82b3
Implement, test IterRetainedStrata
