+++ hstrat/hstrat/__init__.py
@@ -3,5 +3,6 @@
 from ..genome_instrumentation import *  # noqa: F401
 from ..juxtaposition import *  # noqa: F401
 from ..phylogenetic_inference import *  # noqa: F401
+from ..serialization import *  # noqa: F401
 from ..stratum_retention_strategy import *  # noqa: F401
 from ..stratum_retention_viz import *  # noqa: F401
