+++ tests/test_hstrat/test_serialization/test_col_to_dataframe.py
@@ -1,0 +1,28 @@
+import itertools as it
+
+import pytest
+
+from hstrat import hstrat
+
+
+@pytest.mark.parametrize(
+    "retention_policy",
+    [
+        hstrat.perfect_resolution_algo.Policy(),
+        hstrat.nominal_resolution_algo.Policy(),
+        hstrat.fixed_resolution_algo.Policy(fixed_resolution=10),
+    ],
+)
+def test_col_to_dataframe(
+    retention_policy,
+):
+    column = hstrat.HereditaryStratigraphicColumn(
+        stratum_retention_policy=retention_policy,
+    )
+
+    for generation in range(100):
+        column.DepositStratum()
+
+    df = hstrat.col_to_dataframe(column)
+    assert len(df) == column.GetNumStrataRetained()
+    assert df["rank"].max() == column.GetNumStrataDeposited() - 1
