### 2022-12-06 f9e0250
Implement ascii col print, add example
