### 2022-12-06 dbb46d9
Implement ascii col print, add example
