+++ pyproject.toml
@@ -27,6 +27,7 @@ dependencies = [
   "numpy>=1.17.2",
   "mpmath>=1.1.0",
   "opytional>=0.1.0",
+  "prettytable>=3.5.0",
   "python-slugify>=6.1.2",
   "safe_assert>=0.2.0",
   "seaborn>=0.11.2",
@@ -85,6 +86,7 @@ pinned_dependencies_py37 = [
   # pandas not a direct requirement
   # however this pin is necessary due an apparent bug in pip-compile
   "pandas==1.3.5",
+  "prettytable==3.5.0",
   "python-slugify==6.1.2",
   "seaborn==0.12.1",
 ]
@@ -103,6 +105,7 @@ pinned_dependencies_py310 = [
   # pandas not a direct requirement
   # however this pin is necessary due an apparent bug in pip-compile
   "pandas==1.3.5",
+  "prettytable==3.5.0",
   "python-slugify==6.1.2",
   "seaborn==0.12.1",
 ]
