### 2022-12-06 d61142a
Remove bumpversion config for deleted file
