+++ setup.cfg
@@ -3,10 +3,6 @@ current_version = 1.0.5
 commit = True
 tag = True
 
-[bumpversion:file:setup.py]
-search = version="{current_version}"
-replace = version="{new_version}"
-
 [bumpversion:file:hstrat/__init__.py]
 search = __version__ = "{current_version}"
 replace = __version__ = "{new_version}"
