### 2022-12-06 mmore500/hstrat#61 @mmore500
Implemented `hstrat.col_to_ascii` to create a pretty-print representation of the internal state of a column. To support this, I also implemented `hstrat.col_to_dataframe` which creates a pandas DataFrame representation of a hstrat column.

Put together a little demo of `hstrat.col_to_ascii` in `examples/ascii_stratum_retention_viz.py`.

Let me know your thoughts on these improvements.

Apologies for the delayed response! Agreed that this will be a useful tool to peek at what's going on under the hood (and especially to help new users get acquainted).
