### 2022-12-06 mmore500/hstrat#61 @kgd-al
Great! It might not be crucial for everyone but, as my first reaction was 'how to print it?' followed by 'Well let's dig in the code and expose all those private members', it should help all of us print-maniacs.
