### 2022-12-06 f9970ec
Move bumpversion config to own file
Fixes trailing whitespace issue on version bumps
