+++ .bumpversion.cfg
@@ -1,0 +1,16 @@
+[bumpversion]
+current_version = 1.0.5
+commit = True
+tag = True
+
+[bumpversion:file:hstrat/__init__.py]
+search = __version__ = "{current_version}"
+replace = __version__ = "{new_version}"
+
+[bumpversion:file:CITATION.cff]
+search = version: {current_version}
+replace = version: {new_version}
+
+[bumpversion:file:pyproject.toml]
+search = version = "{current_version}"
+replace = version = "{new_version}"
