### 2022-12-06 757692b
Bump version: 1.0.5 → 1.0.6
