+++ pyproject.toml
@@ -43,7 +43,7 @@ keywords = [
   "hstrat",
 ]
 requires-python = ">=3.7"
-version = "1.0.5"
+version = "1.0.6"
 
 [project.urls]
 homepage = "https://github.com/mmore500/hstrat"
