### 2022-12-06 c20ea1d
Bump version: 1.0.5 → 1.0.6
