### 2022-12-06 f5bb1c3
Revert "Move bumpversion config to own file"
This reverts commit f9970eceb9c209134179ecea77b80681dacebb13.
