### 2022-12-07 mmore500/hstrat#67 @osorensen
Editorial comments on JOSS submission
I'll post editorial comments on the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4866) as issues in this repository. This is the first one.

The name "Matthew A. Moreno" is inconsistently spelled you `.bib` file, causing the references to not appear as they should. See, e.g., on line 16.

The source of the problem seems to be here: https://github.com/mmore500/hstrat/blob/f5bb1c3cd817b299d6336d35ad7d44e94ce5e789/joss/paper.bib#L370-L409
