### 2022-12-07 mmore500/hstrat#68 @osorensen
Editorial comments on JOSS submission
Another comment on [the JOSS submission](https://github.com/openjournals/joss-reviews/issues/4866).

On lines 150-151 you write the following: "Indeed, at a minimum, we plan to update the Python library to include support for wrapped these native components [@pybind11]."

I don't fully understand this sentence. Would it make sense to replace the word "wrapped" with "wrapping", so it becomes "Indeed, at a minimum, we plan to update the Python library to include support for **wrapping** these native components [@pybind11]."
