### 2022-12-07 mmore500/hstrat#69 @osorensen
Non-existent DOI
This is an editorial comment on [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4866).

The DOI for the paper Ackley and Small (2014) does not seem to exist. I see that the DOI in your `.bib` is identical to the DOI reported at the MIT Press website (https://direct.mit.edu/isal/proceedings/alife2014/606/99046), but clicking on it I get the message that it does not exist (https://doi.org/10.1162/978-0-262-32621-6-ch098).

Could you please fix this be removing the DOI entry here?

https://github.com/mmore500/hstrat/blob/f5bb1c3cd817b299d6336d35ad7d44e94ce5e789/joss/paper.bib#L20-L31
