### 2022-12-07 mmore500/hstrat#70 @osorensen
Non-existent DOI
Here's another editorial comment on the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4866).

Just like in #69, to DOI for Bohm and Hintze (2017) does not exist, even though the one in your bib file is identical to what is reported on MIT Press (https://direct.mit.edu/isal/proceedings/ecal2017/76/99520).

Could you please fix by removing the DOI entry here?

https://github.com/mmore500/hstrat/blob/f5bb1c3cd817b299d6336d35ad7d44e94ce5e789/joss/paper.bib#L103-L114
