### 2022-12-07 mmore500/hstrat#71 @osorensen
Reference formatting
This is an editorial comment on the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4866).

Could you please protect the capital letter in Geb by adding a set of curly braces? That is, change `title = {Maximum Individual Complexity is Indefinitely Scalable in Geb}, ` to `title = {Maximum Individual Complexity is Indefinitely Scalable in {Geb}}, `

https://github.com/mmore500/hstrat/blob/f5bb1c3cd817b299d6336d35ad7d44e94ce5e789/joss/paper.bib#L139-L152
