### 2022-12-07 mmore500/hstrat#73 @osorensen
Reference formatting
This is an editorial comment on the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4866).

Could you please protect the uppercase letters in Python and C++ in the reference below, cf. #71 and #72?

https://github.com/mmore500/hstrat/blob/f5bb1c3cd817b299d6336d35ad7d44e94ce5e789/joss/paper.bib#L473-L478
