### 2022-12-07 mmore500/hstrat#75 @osorensen
Reference formatting
This is an editorial comment on the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4866).

Please protect the capital E in "Evolving" in this reference:

https://github.com/mmore500/hstrat/blob/f5bb1c3cd817b299d6336d35ad7d44e94ce5e789/joss/paper.bib#L358-L369

Alternatively, add `chapter={15}` as a separate key, and change to `title = {Evolving Deep Neural Networks}`
