### 2022-12-07 mmore500/hstrat#76 @osorensen
Reference formatting
Editorial comment on [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4866)

Please keep OSF all capital letters by changing `{Osf}` to `{{OSF}}`.

https://github.com/mmore500/hstrat/blob/f5bb1c3cd817b299d6336d35ad7d44e94ce5e789/joss/paper.bib#L370-L378
