### 2022-12-07 mmore500/hstrat#77 @osorensen
Link pointing to the wrong repository?
This is an editorial comment on the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4866)

The link https://doi.org/10.5281/zenodo.4141943 for the reference Ofria et al. (2020) leads me to the following webpage. Is that correct?

![image](https://user-images.githubusercontent.com/21175639/206179789-dbd2b031-1b92-4548-afa3-502207828f10.png)
