### 2022-12-08 mmore500/hstrat#77 @mmore500
Yes. That is the intended software citation.
