+++ joss/paper.bib
@@ -372,7 +372,7 @@
   author = {Moreno, Matthew A and Ofria, Charles},
   year = 2020,
   month = {Jun},
-  publisher = {Osf},
+  publisher = {{OSF}},
   doi = {10.17605/osf.io/53vgh},
   url = {osf.io/53vgh},
 }
