+++ joss/paper.bib
@@ -356,12 +356,13 @@
   eprint = {https://academic.oup.com/mit-press-scholarship-online/book/0/chapter/182250754/chapter-ag-pdf/44683482/book\_22123\_section\_182250754.ag.pdf},
 }
 @incollection{miikkulainen2019evolving,
-  title = {Chapter 15 - Evolving Deep Neural Networks},
+  title = {Evolving Deep Neural Networks},
   author = {Risto Miikkulainen and Jason Liang and Elliot Meyerson and Aditya Rawal and Daniel Fink and Olivier Francon and Bala Raju and Hormoz Shahrzad and Arshak Navruzyan and Nigel Duffy and Babak Hodjat},
   year = 2019,
   booktitle = {Artificial Intelligence in the Age of Neural Networks and Brain Computing},
   publisher = {Academic Press},
   pages = {293--312},
+  chapter = {15},
   doi = {10.1016/b978-0-12-815480-9.00015-3},
   isbn = {978-0-12-815480-9},
   url = {https://www.sciencedirect.com/science/article/pii/B9780128154809000153},
