+++ joss/paper.bib
@@ -344,7 +344,7 @@
   doi = {10.1109/mc.2008.17},
 }
 @incollection{mcshea2011miscellaneous,
-  title = {19The Miscellaneous Transitions in Evolution},
+  title = {The Miscellaneous Transitions in Evolution},
   author = {McShea, Daniel W. and Simpson, Carl},
   year = 2011,
   month = {04},
