+++ joss/paper.bib
@@ -472,7 +472,7 @@
   lccn = 2016462290,
 }
 @misc{pybind11,
-  title = {pybind11 -- Seamless operability between C++11 and Python},
+  title = {pybind11 -- Seamless operability between {C}++11 and {P}ython},
   author = {Wenzel Jakob and Jason Rhinelander and Dean Moldovan},
   year = 2017,
   note = {https://github.com/pybind/pybind11},
