+++ joss/paper.bib
@@ -163,7 +163,7 @@
   eprint = {https://royalsocietypublishing.org/doi/pdf/10.1098/rspb.2010.1557},
 }
 @article{cock2009biopython,
-  title = {Biopython: freely available Python tools for computational molecular biology and bioinformatics},
+  title = {Biopython: freely available {P}ython tools for computational molecular biology and bioinformatics},
   author = {Cock, Peter J. A. and Antao, Tiago and Chang, Jeffrey T. and Chapman, Brad A. and Cox, Cymon J. and Dalke, Andrew and Friedberg, Iddo and Hamelryck, Thomas and Kauff, Frank and Wilczynski, Bartek and de Hoon, Michiel J. L.},
   year = 2009,
   month = {03},
