+++ joss/paper.bib
@@ -137,7 +137,7 @@
   doi = {10.1109/cec.2003.1299834},
 }
 @article{channon2019maximum,
-  title = {Maximum Individual Complexity is Indefinitely Scalable in Geb},
+  title = {Maximum Individual Complexity is Indefinitely Scalable in {Geb}},
   author = {Channon, Alastair},
   year = 2019,
   month = {05},
