### 2022-12-08 aad51db
Protect SWIG
