+++ joss/paper.bib
@@ -64,7 +64,7 @@
   isbn = {155860510x},
 }
 @article{beazley2003automated,
-  title = {Automated scientific software scripting with SWIG},
+  title = {Automated scientific software scripting with {SWIG}},
   author = {D.M. Beazley},
   year = 2003,
   journal = {Future Generation Computer Systems},
