### 2022-12-08 5d480b8
Protect MABE, DEAP, SpiNNaker
