+++ joss/paper.bib
@@ -101,7 +101,7 @@
 }
 % dead doi
 @inproceedings{bohm2017mabe,
-  title = {MABE (Modular Agent Based Evolver): A framework for digital evolution research},
+  title = {{MABE} (Modular Agent Based Evolver): A framework for digital evolution research},
   author = {Bohm, Clifford and G., Nitash C. and Hintze, Arend},
   year = 2017,
   month = {09},
@@ -113,7 +113,7 @@
   eprint = {https://direct.mit.edu/isal/proceedings-pdf/ecal2017/29/76/1904506/isal\_a\_016.pdf},
 }
 @inproceedings{bohm2019mabe,
-  title = {MABE 2.0: An Introduction to MABE and a Road Map for the Future of MABE Development},
+  title = {{MABE} 2.0: An Introduction to MABE and a Road Map for the Future of MABE Development},
   author = {Bohm, Clifford and Lalejini, Alexander and Schossau, Jory and Ofria, Charles},
   year = 2019,
   booktitle = {Proceedings of the Genetic and Evolutionary Computation Conference Companion},
@@ -177,7 +177,7 @@
   eprint = {https://academic.oup.com/bioinformatics/article-pdf/25/11/1422/944180/btp163.pdf},
 }
 @inproceedings{de2012deap,
-  title = {DEAP: A Python Framework for Evolutionary Algorithms},
+  title = {{DEAP}: A Python Framework for Evolutionary Algorithms},
   author = {De Rainville, Fran\c{c}ois-Michel and Fortin, F\'{e}lix-Antoine and Gardner, Marc-Andr\'{e} and Parizeau, Marc and Gagn\'{e}, Christian},
   year = 2012,
   booktitle = {Proceedings of the 14th Annual Conference Companion on Genetic and Evolutionary Computation},
@@ -217,7 +217,7 @@
   edition = 2,
 }
 @article{furber2014spinnaker,
-  title = {The SpiNNaker Project},
+  title = {The {SpiNNaker} Project},
   author = {Furber, Steve B. and Galluppi, Francesco and Temple, Steve and Plana, Luis A.},
   year = 2014,
   journal = {Proceedings of the IEEE},
