+++ joss/paper.bib
@@ -441,7 +441,7 @@
   doi = {10.1162/106454604773563612},
 }
 @software{ofria2020empirical,
-  title = {Empirical: C++ library for efficient, reliable, and accessible scientific software},
+  title = {{Empirical}: {C++} library for efficient, reliable, and accessible scientific software},
   author = {Ofria, Charles and Moreno, Matthew Andres and Dolson, Emily and Lalejini, Alex and Rodriguez Papa, Santiago and Fenton, Jake and Perry, Katherine and Jorgensen, Steven and hoffmanriley and grenewode and Baldwin Edwards, Oliver and Stredwick, Jason and cgnitash and theycallmeHeem and Vostinar, Anya and Moreno, Ryan and Schossau, Jory and Zaman, Luis and djrain},
   year = 2020,
   month = 10,
