### 2022-12-08 bf5dd7e
Protect more MABE
