+++ joss/paper.bib
@@ -113,7 +113,7 @@
   eprint = {https://direct.mit.edu/isal/proceedings-pdf/ecal2017/29/76/1904506/isal\_a\_016.pdf},
 }
 @inproceedings{bohm2019mabe,
-  title = {{MABE} 2.0: An Introduction to MABE and a Road Map for the Future of MABE Development},
+  title = {{MABE} 2.0: An Introduction to {MABE} and a Road Map for the Future of {MABE} Development},
   author = {Bohm, Clifford and Lalejini, Alexander and Schossau, Jory and Ofria, Charles},
   year = 2019,
   booktitle = {Proceedings of the Genetic and Evolutionary Computation Conference Companion},
