### 2022-12-08 6939845
Fixup more bibtex protection
