+++ joss/paper.bib
@@ -163,7 +163,7 @@
   eprint = {https://royalsocietypublishing.org/doi/pdf/10.1098/rspb.2010.1557},
 }
 @article{cock2009biopython,
-  title = {Biopython: freely available {P}ython tools for computational molecular biology and bioinformatics},
+  title = {Biopython: freely available {Python} tools for computational molecular biology and bioinformatics},
   author = {Cock, Peter J. A. and Antao, Tiago and Chang, Jeffrey T. and Chapman, Brad A. and Cox, Cymon J. and Dalke, Andrew and Friedberg, Iddo and Hamelryck, Thomas and Kauff, Frank and Wilczynski, Bartek and de Hoon, Michiel J. L.},
   year = 2009,
   month = {03},
@@ -177,7 +177,7 @@
   eprint = {https://academic.oup.com/bioinformatics/article-pdf/25/11/1422/944180/btp163.pdf},
 }
 @inproceedings{de2012deap,
-  title = {{DEAP}: A Python Framework for Evolutionary Algorithms},
+  title = {{DEAP}: A {Python} Framework for Evolutionary Algorithms},
   author = {De Rainville, Fran\c{c}ois-Michel and Fortin, F\'{e}lix-Antoine and Gardner, Marc-Andr\'{e} and Parizeau, Marc and Gagn\'{e}, Christian},
   year = 2012,
   booktitle = {Proceedings of the 14th Annual Conference Companion on Genetic and Evolutionary Computation},
@@ -236,7 +236,7 @@
   doi = {10.7312/crac92306-005},
 }
 @inproceedings{godin2019apoget,
-  title = {APOGeT: Automated Phylogeny Over Geological Timescales},
+  title = {{APOGeT}: Automated Phylogeny Over Geological Timescales},
   author = {Godin-Dubois, Kevin and Cussat-Blanc, Sylvain and Duthen, Yves},
   year = 2019,
   month = {08},
@@ -313,7 +313,7 @@
   eprint = {https://direct.mit.edu/isal/proceedings-pdf/isal2019/31/388/1903469/isal\_a\_00191.pdf},
 }
 @article{lauterbach2021path,
-  title = {The Path to Successful Wafer-Scale Integration: The Cerebras Story},
+  title = {The Path to Successful Wafer-Scale Integration: The {Cerebras} Story},
   author = {Lauterbach, Gary},
   year = 2021,
   journal = {IEEE Micro},
@@ -472,7 +472,7 @@
   lccn = 2016462290,
 }
 @misc{pybind11,
-  title = {pybind11 -- Seamless operability between {C}++11 and {P}ython},
+  title = {{pybind11} -- Seamless operability between {C++11} and {Python}},
   author = {Wenzel Jakob and Jason Rhinelander and Dean Moldovan},
   year = 2017,
   note = {https://github.com/pybind/pybind11},
