### 2022-12-08 cd83d1b
Disable dead DOI
Closes #70
