### 2022-12-08 64dc8cb
Disable dead DOI
Closes #70
