+++ joss/paper.bib
@@ -108,7 +108,7 @@
   series = {ALIFE 2022: The 2022 Conference on Artificial Life},
   volume = {ECAL 2017, the Fourteenth European Conference on Artificial Life},
   pages = {76--83},
-  doi = {10.1162/isal_a_016},
+  % doi = {10.1162/isal_a_016},
   url = {https://direct.mit.edu/isal/proceedings/ecal2017/76/99520},
   eprint = {https://direct.mit.edu/isal/proceedings-pdf/ecal2017/29/76/1904506/isal\_a\_016.pdf},
 }
