### 2022-12-08 e0d2003
Fix typo
Closes #68
