### 2022-12-08 3006fbf
Fix typo
Closes #68
