### 2022-12-08 9057cb8
Disable dead DOI
Closes #69
