### 2022-12-08 9796446
Disable dead DOI
Closes #69
