+++ joss/paper.bib
@@ -25,7 +25,7 @@
   series = {ALIFE 2022: The 2022 Conference on Artificial Life},
   volume = {ALIFE 14: The Fourteenth International Conference on the Synthesis and Simulation of Living Systems},
   pages = {606--613},
-  doi = {10.1162/978-0-262-32621-6-ch098},
+  % doi = {10.1162/978-0-262-32621-6-ch098},
   url = {https://direct.mit.edu/isal/proceedings/alife2014/606/99046},
   eprint = {https://direct.mit.edu/isal/proceedings-pdf/alife2014/26/606/1901859/978-0-262-32621-6-ch098.pdf},
 }
