+++ joss/paper.bib
@@ -370,7 +370,7 @@
 }
 @misc{moreno2020practical,
   title = {Practical Steps Toward Indefinite Scalability: In Pursuit of Robust Computational Substrates for Open-Ended Evolution},
-  author = {Moreno, Matthew A and Ofria, Charles},
+  author = {Moreno, Matthew Andres and Ofria, Charles},
   year = 2020,
   month = {Jun},
   publisher = {{OSF}},
@@ -379,7 +379,7 @@
 }
 @inproceedings{moreno2021case,
   title = {Case Study of Novelty, Complexity, and Adaptation in a Multicellular System},
-  author = {Moreno, Matthew A and Rodriguez Papa, Santiago and Ofria, Charles},
+  author = {Moreno, Matthew Andres and Rodriguez Papa, Santiago and Ofria, Charles},
   year = 2021,
   booktitle = {OEE4: The Fourth Workshop on Open-Ended Evolution},
   maintitle = {The 2021 Conference on Artificial Life},
