### 2022-12-09 f168018
Bump version: 1.0.6 → 1.1.0
