### 2022-12-09 dd43397
Bump version: 1.0.6 → 1.1.0
