+++ hstrat/__init__.py
@@ -2,7 +2,7 @@
 
 __author__ = """Matthew Andres Moreno"""
 __email__ = "m.more500@gmail.com"
-__version__ = "1.0.6"
+__version__ = "1.1.0"
 
 from . import (
     _auxiliary_lib,
