+++ CITATION.cff
@@ -11,7 +11,7 @@ authors:
     given-names: Charles
     orcid: https://orcid.org/0000-0003-2924-1732
 title: "hstrat"
-version: 1.0.6
+version: 1.1.0
 preferred-citation:
   type: article
   authors:
