+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 1.0.6
+current_version = 1.1.0
 commit = True
 tag = True
 
@@ -26,3 +26,4 @@ test = pytest
 
 [tool:pytest]
 collect_ignore = ['setup.py']
+
