+++ pyproject.toml
@@ -43,7 +43,7 @@ keywords = [
   "hstrat",
 ]
 requires-python = ">=3.7"
-version = "1.0.6"
+version = "1.1.0"
 
 [project.urls]
 homepage = "https://github.com/mmore500/hstrat"
