### 2022-12-09 #4866 @osorensen
Thanks @mmore500.

Could you please make sure the title of the Zenodo archive exactly matches that of the paper: "hstrat: a Python Package for phylogenetic inference on distributed digital evolution populations", and let me know when it's done?
