### 2022-12-09 #4866 @mmore500
Ah, Looks like I had clicked save instead of publish so the title change wasn’t applied publicly. I believe the correct title should be showing up now?
