+++ joss/paper.bib
@@ -23,7 +23,7 @@
   year = 2014,
   month = {07},
   series = {ALIFE 2022: The 2022 Conference on Artificial Life},
-  volume = {ALIFE 14: The Fourteenth International Conference on the Synthesis and Simulation of Living Systems},
+  volume = {14},
   pages = {606--613},
   % doi = {10.1162/978-0-262-32621-6-ch098},
   url = {https://direct.mit.edu/isal/proceedings/alife2014/606/99046},
