+++ joss/paper.bib
@@ -23,7 +23,7 @@
   year = 2014,
   month = {07},
   series = {ALIFE 2022: The 2022 Conference on Artificial Life},
-  volume = {14},
+  volume = 14,
   pages = {606--613},
   % doi = {10.1162/978-0-262-32621-6-ch098},
   url = {https://direct.mit.edu/isal/proceedings/alife2014/606/99046},
@@ -105,8 +105,8 @@
   author = {Bohm, Clifford and G., Nitash C. and Hintze, Arend},
   year = 2017,
   month = {09},
-  series = {ALIFE 2022: The 2022 Conference on Artificial Life},
-  volume = {ECAL 2017, the Fourteenth European Conference on Artificial Life},
+  series = {ECAl, the European Conference on Artificial Life},
+  volume = 14,
   pages = {76--83},
   % doi = {10.1162/isal_a_016},
   url = {https://direct.mit.edu/isal/proceedings/ecal2017/76/99520},
@@ -305,8 +305,7 @@
   author = {Langdon, William B. and Banzhaf, Wolfgang},
   year = 2019,
   month = {07},
-  series = {ALIFE 2022: The 2022 Conference on Artificial Life},
-  volume = {ALIFE 2019: The 2019 Conference on Artificial Life},
+  series = {The Conference on Artificial Life},
   pages = {388--395},
   doi = {10.1162/isal_a_00191},
   url = {https://doi.org/10.1162/isal\_a\_00191},
@@ -400,8 +399,7 @@
   year = 2022,
   month = {07},
   booktitle = {Proceedings of the Genetic and Evolutionary Computation Conference Companion},
-  series = {ALIFE 2022: The 2022 Conference on Artificial Life},
-  volume = {ALIFE 2022: The 2022 Conference on Artificial Life},
+  series = {The Conference on Artificial Life},
   pages = {65--66},
   doi = {10.1162/isal_a_00550},
   url = {https://doi.org/10.1162/isal\_a\_00550},
