+++ joss/paper.bib
@@ -72,7 +72,6 @@
   number = 5,
   pages = {599--609},
   doi = {10.1016/s0167-739x(02)00171-1},
-  issn = {0167-739x},
   url = {https://www.sciencedirect.com/science/article/pii/S0167739X02001711},
   note = {Tools for Program Development and Analysis. Best papers from two Technical Sessions, at ICCS2001, San Francisco, CA, USA, and ICCS2002, Amsterdam, The Netherlands},
 }
@@ -390,7 +389,6 @@
   journal = {Frontiers in Ecology and Evolution},
   volume = 10,
   doi = {10.3389/fevo.2022.750837},
-  issn = {2296-701x},
   url = {https://www.frontiersin.org/articles/10.3389/fevo.2022.750837},
 }
 @inproceedings{moreno2022hereditary,
