### 2022-12-16 67a5347
Try patching @editorialbot parse errors
Move commented out dois between entries, simplify series names
