+++ joss/paper.bib
@@ -17,15 +17,15 @@
   url = {https://www.usenix.org/conference/hotosxiii/pursue-robust-indefinite-scalability},
 }
 % dead doi
+% doi = {10.1162/978-0-262-32621-6-ch098},
 @inproceedings{ackley2014indefinitely,
   title = {Indefinitely Scalable Computing = Artificial Life Engineering},
   author = {Ackley, David and Small, Trent},
   year = 2014,
   month = {07},
-  series = {ALIFE 2022: The 2022 Conference on Artificial Life},
+  series = {The Conference on Artificial Life},
   volume = 14,
   pages = {606--613},
-  % doi = {10.1162/978-0-262-32621-6-ch098},
   url = {https://direct.mit.edu/isal/proceedings/alife2014/606/99046},
   eprint = {https://direct.mit.edu/isal/proceedings-pdf/alife2014/26/606/1901859/978-0-262-32621-6-ch098.pdf},
 }
@@ -99,15 +99,15 @@
   note = {US Patent App. 12/267,287},
 }
 % dead doi
+% doi = {10.1162/isal_a_016},
 @inproceedings{bohm2017mabe,
   title = {{MABE} (Modular Agent Based Evolver): A framework for digital evolution research},
   author = {Bohm, Clifford and G., Nitash C. and Hintze, Arend},
   year = 2017,
   month = {09},
-  series = {ECAl, the European Conference on Artificial Life},
+  series = {ECAl the European Conference on Artificial Life},
   volume = 14,
   pages = {76--83},
-  % doi = {10.1162/isal_a_016},
   url = {https://direct.mit.edu/isal/proceedings/ecal2017/76/99520},
   eprint = {https://direct.mit.edu/isal/proceedings-pdf/ecal2017/29/76/1904506/isal\_a\_016.pdf},
 }
