+++ joss/paper.bib
@@ -16,8 +16,6 @@
   address = {Napa, CA},
   url = {https://www.usenix.org/conference/hotosxiii/pursue-robust-indefinite-scalability},
 }
-% dead doi
-% doi = {10.1162/978-0-262-32621-6-ch098},
 @inproceedings{ackley2014indefinitely,
   title = {Indefinitely Scalable Computing = Artificial Life Engineering},
   author = {Ackley, David and Small, Trent},
@@ -26,6 +24,7 @@
   series = {The Conference on Artificial Life},
   volume = 14,
   pages = {606--613},
+  doi = {10.7551/978-0-262-32621-6-ch098},
   url = {https://direct.mit.edu/isal/proceedings/alife2014/606/99046},
   eprint = {https://direct.mit.edu/isal/proceedings-pdf/alife2014/26/606/1901859/978-0-262-32621-6-ch098.pdf},
 }
@@ -98,8 +97,6 @@
   publisher = {Google Patents},
   note = {US Patent App. 12/267,287},
 }
-% dead doi
-% doi = {10.1162/isal_a_016},
 @inproceedings{bohm2017mabe,
   title = {{MABE} (Modular Agent Based Evolver): A framework for digital evolution research},
   author = {Bohm, Clifford and G., Nitash C. and Hintze, Arend},
@@ -108,6 +105,7 @@
   series = {ECAl the European Conference on Artificial Life},
   volume = 14,
   pages = {76--83},
+  doi = {10.7551/ecal_a_016},
   url = {https://direct.mit.edu/isal/proceedings/ecal2017/76/99520},
   eprint = {https://direct.mit.edu/isal/proceedings-pdf/ecal2017/29/76/1904506/isal\_a\_016.pdf},
 }
