### 2022-12-16 #4866 @mmore500
@osorensen think I have the bib patched up! Looks like the two dois that had been dead earlier have been revived, so I added them back in.
